Alterado a Estrutura dos Enums do Projeto
