--- backend/thebookclub/src/main/java/com/projectg2/thebookclub/model/FormaEnvio.java
@@ -1,8 +1,0 @@
-package br.com.digitalhouse.thebookclub.modelo;
-
-public enum FormaEnvio {
-	
-	CORREIOS, 
-	RETIRADANALOJA,
-	TRANSPORTADORA;
-}
