--- backend/thebookclub/src/main/java/com/projectg2/thebookclub/model/StatusPedido.java
@@ -1,11 +1,0 @@
-package br.com.digitalhouse.thebookclub.modelo;
-
-public enum StatusPedido {
-	
-	AGUARDANDO_PAGAMENTO,
-	PROCESSANDO,
-	PAGTO_APROVADO,
-	PAGTO_RECUSADO,
-	ENVIADO,
-	ENTREGUE;
-}
