--- backend/thebookclub/src/main/java/com/projectg2/thebookclub/model/TipoPagamento.java
@@ -1,11 +1,0 @@
-package br.com.digitalhouse.thebookclub.modelo;
-
-public enum TipoPagamento {
-	
-	CREDITO,
-	DEBITO,
-	PIX,
-	BOLETO,
-	TRANSFERBANCARIA;
-	
-}
